docs-requirements.txt
-e .
sphinx==1.6.5
sphinxcontrib-zopeext==0.2.1
